community: add kata-micro-kernel in sigs.yaml
reason: Kata-micro-kernel is a micro kernel built by specific kernel config.
It is used by kata-containers when starting a secure container.
The guest kernel is passed to the hypervisor and used to boot the virtual machine. The default kernel provided in Kata Containers is highly optimized for kernel boot time and minimal memory footprint, providing only those services required by a container workload.

Signed-off-by: yangfeiyu <yangfeiyu20102011@163.com>