New package perl-DBD-MariaDB introduction
perl-DBD-MariaDB package summary:
MariaDB and MySQL driver for the Perl5 Database Interface (DBI)

The reason for this proposal:
Due to mysqlhotcopy needs DBI/DBD support. And it implements with perl interface, which is named "perl(DBD::MariaDB)".
So for fitting the full MariaDB/Mysql functionality, we need to support its hotcopy function.