[DB SIG] Introduce livy into openEuler
Apache Livy is an open source REST interface for interacting with Apache Spark
from anywhere. It supports executing snippets of code or programs in a Spark
context that runs locally or in Apache Hadoop YARN.

At the same time, it's a famous connection and widely use in bigdata DB solution,
between DB softwares and Bigdata spark clusters.
For making sure the good maintainence, we think the good way to set the maintaining
resposibilty to DB SIG.