libmetal: move libmetal repo from sig-ai-bigdata to sig-embedded.
OpenAMP project is a framework of cross-OS/runtime communication.
It's mainly designed for embedded system.

Libmetal is a part of OpenAMP project, and is maintained and released
together with libopenamp.

Since 22.03, OpenAMP project will be introdued in openEuler Embedded and
maintained by sig-embedded.

So it's better to put OpenAMP project related software package together.

It seems that libmetal now is not used in sig-ai-bigdata, no new branch
that goes with openEuler release, and no update to the latest version.

Therefore, this commit is to move libmetal repo from sig-ai-bigdata to
sig-embedded for better maintainence.

Signed-off-by: Wayne Ren <renwei41@huawei.com>