update A-Tune Sig Maintainers
Due to job changes, the following people will no longer
serve as A-Tune Sig Maintainers. We are very grateful for
their contributions to A-Tune Sig over the past few years.
The following persons are no longer A-Tune Sig Maintainers:
Zhipeng Xie
Cihua Li
Ruoshu Gao

At the same time, I would like to recommend Dr. Chen Donghui
to become the A-Tune SIG Maintainer

The outdated README.md is also updated in this PR

Signed-off-by: Zhipeng Xie <xiezhipeng1@huawei.com>